requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
alembic==1.8.1
    # via apache-airflow
anyio==3.6.1
    # via httpcore
apache-airflow==2.3.3
    # via
    #   -r requirements.in
    #   apache-airflow-providers-postgres
apache-airflow-providers-common-sql==1.1.0
    # via apache-airflow-providers-sqlite
apache-airflow-providers-ftp==3.1.0
    # via apache-airflow
apache-airflow-providers-http==4.0.0
    # via apache-airflow
apache-airflow-providers-imap==3.0.0
    # via apache-airflow
apache-airflow-providers-postgres==2.3.0
    # via -r requirements.in
apache-airflow-providers-sqlite==3.2.0
    # via apache-airflow
apispec[yaml]==3.3.2
    # via flask-appbuilder
argcomplete==1.12.2
    # via apache-airflow
attrs==20.3.0
    # via
    #   apache-airflow
    #   cattrs
    #   jsonschema
babel==2.10.3
    # via flask-babel
backports-zoneinfo==0.2.1
    # via psycopg
blinker==1.5
    # via apache-airflow
cachelib==0.9.0
    # via
    #   flask-caching
    #   flask-session
cattrs==1.10.0
    # via apache-airflow
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   clickclick
    #   flask
    #   flask-appbuilder
clickclick==20.10.2
    # via connexion
colorama==0.4.5
    # via flask-appbuilder
colorlog==4.8.0
    # via apache-airflow
commonmark==0.9.1
    # via rich
connexion[flask,swagger-ui]==2.14.1
    # via apache-airflow
cron-descriptor==1.2.31
    # via apache-airflow
croniter==1.3.7
    # via apache-airflow
cryptography==38.0.1
    # via apache-airflow
deprecated==1.2.13
    # via apache-airflow
dill==0.3.5.1
    # via apache-airflow
dnspython==2.2.1
    # via email-validator
docutils==0.19
    # via python-daemon
email-validator==1.2.1
    # via flask-appbuilder
flask==2.2.2
    # via
    #   apache-airflow
    #   connexion
    #   flask-appbuilder
    #   flask-babel
    #   flask-caching
    #   flask-jwt-extended
    #   flask-login
    #   flask-session
    #   flask-sqlalchemy
    #   flask-wtf
flask-appbuilder==4.1.2
    # via apache-airflow
flask-babel==2.0.0
    # via flask-appbuilder
flask-caching==2.0.1
    # via apache-airflow
flask-jwt-extended==4.4.4
    # via flask-appbuilder
flask-login==0.6.2
    # via
    #   apache-airflow
    #   flask-appbuilder
flask-session==0.4.0
    # via apache-airflow
flask-sqlalchemy==2.5.1
    # via flask-appbuilder
flask-wtf==0.15.1
    # via
    #   apache-airflow
    #   flask-appbuilder
graphviz==0.20.1
    # via apache-airflow
gunicorn==20.1.0
    # via apache-airflow
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via apache-airflow
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
    #   rfc3986
importlib-metadata==4.12.0
    # via
    #   alembic
    #   apache-airflow
    #   flask
    #   markdown
importlib-resources==5.9.0
    # via
    #   alembic
    #   apache-airflow
    #   jsonschema
inflection==0.5.1
    # via connexion
itsdangerous==2.1.2
    # via
    #   apache-airflow
    #   connexion
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via
    #   apache-airflow
    #   flask
    #   flask-babel
    #   python-nvd3
    #   swagger-ui-bundle
jsonschema==4.15.0
    # via
    #   apache-airflow
    #   connexion
    #   flask-appbuilder
lazy-object-proxy==1.7.1
    # via apache-airflow
linkify-it-py==2.0.0
    # via apache-airflow
lockfile==0.12.2
    # via
    #   apache-airflow
    #   python-daemon
mako==1.2.2
    # via alembic
markdown==3.4.1
    # via apache-airflow
markdown-it-py==2.1.0
    # via
    #   apache-airflow
    #   mdit-py-plugins
markupsafe==2.1.1
    # via
    #   apache-airflow
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
marshmallow==3.17.1
    # via
    #   flask-appbuilder
    #   marshmallow-enum
    #   marshmallow-oneofschema
    #   marshmallow-sqlalchemy
marshmallow-enum==1.5.1
    # via flask-appbuilder
marshmallow-oneofschema==3.0.1
    # via apache-airflow
marshmallow-sqlalchemy==0.26.1
    # via flask-appbuilder
mdit-py-plugins==0.3.0
    # via apache-airflow
mdurl==0.1.2
    # via markdown-it-py
packaging==21.3
    # via
    #   apache-airflow
    #   connexion
    #   marshmallow
pathspec==0.9.0
    # via apache-airflow
pendulum==2.1.2
    # via apache-airflow
pkgutil-resolve-name==1.3.10
    # via jsonschema
pluggy==1.0.0
    # via apache-airflow
prison==0.2.1
    # via flask-appbuilder
psutil==5.9.2
    # via apache-airflow
psycopg==3.1.1
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via apache-airflow-providers-postgres
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via
    #   apache-airflow
    #   rich
pyjwt==2.4.0
    # via
    #   apache-airflow
    #   flask-appbuilder
    #   flask-jwt-extended
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-daemon==2.3.1
    # via apache-airflow
python-dateutil==2.8.2
    # via
    #   apache-airflow
    #   croniter
    #   flask-appbuilder
    #   pendulum
python-nvd3==0.15.0
    # via apache-airflow
python-slugify==6.1.2
    # via
    #   apache-airflow
    #   python-nvd3
pytz==2022.2.1
    # via
    #   babel
    #   flask-babel
pytzdata==2020.1
    # via pendulum
pyyaml==6.0
    # via
    #   apispec
    #   clickclick
    #   connexion
requests==2.28.1
    # via
    #   apache-airflow-providers-http
    #   connexion
    #   requests-toolbelt
requests-toolbelt==0.9.1
    # via apache-airflow-providers-http
rfc3986[idna2008]==1.5.0
    # via httpx
rich==12.5.1
    # via apache-airflow
setproctitle==1.3.2
    # via apache-airflow
six==1.16.0
    # via
    #   prison
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.41
    # via
    #   alembic
    #   apache-airflow
    #   flask-appbuilder
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   sqlalchemy-jsonfield
    #   sqlalchemy-utils
sqlalchemy-jsonfield==1.0.0
    # via apache-airflow
sqlalchemy-utils==0.38.3
    # via flask-appbuilder
sqlparse==0.4.2
    # via apache-airflow-providers-common-sql
swagger-ui-bundle==0.0.9
    # via connexion
tabulate==0.8.10
    # via apache-airflow
tenacity==8.0.1
    # via apache-airflow
termcolor==1.1.0
    # via apache-airflow
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.3.0
    # via
    #   apache-airflow
    #   psycopg
    #   rich
uc-micro-py==1.0.1
    # via linkify-it-py
unicodecsv==0.14.1
    # via apache-airflow
urllib3==1.26.12
    # via requests
werkzeug==2.2.2
    # via
    #   apache-airflow
    #   connexion
    #   flask
    #   flask-jwt-extended
    #   flask-login
wrapt==1.14.1
    # via deprecated
wtforms==2.3.3
    # via
    #   flask-appbuilder
    #   flask-wtf
zipp==3.8.1
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
